+++ src/components/vCard.vue
@@ -2,9 +2,7 @@
 import { type Item } from '../types'
 import vBtn from './vBtn.vue'
 
-defineProps({
-  item: Object as () => Item,
-})
+defineProps<{ item: Item }>()
 
 const openUrl = (url: string | undefined) => {
   if (url) {
@@ -15,22 +13,22 @@ const openUrl = (url: string | undefined) => {
 
 <template>
   <div
-    class="group relative rounded-xl bg-foreground/5 border border-foreground/10 backdrop-blur-xl hover:bg-foreground/10 ta hover:scale-[.99] ta shadow-sm flex flex-col"
+    class="group relative flex flex-col rounded-xl bg-foreground/5 p-1 shadow-xs backdrop-blur-lg ta-200 hover:scale-[.98] hover:bg-foreground/15 hover:backdrop-blur-xl"
   >
     <div class="flex flex-col gap-y-1.5 p-6">
       <RouterLink
         :to="`/item/${item?.id}`"
-        class="text-2xl font-bold leading-none tracking-tight"
+        class="text-2xl leading-none font-bold tracking-tight"
         :title="'ID: ' + item?.id"
       >
         {{ item?.name }}
       </RouterLink>
       <p class="text-base">{{ item?.category }}</p>
     </div>
-    <div class="p-4 pt-0 flex-grow">
-      <div class="flex flex-wrap gap-2 mb-4">
+    <div class="flex-grow p-4 pt-0">
+      <div class="mb-4 flex flex-wrap gap-2">
         <RouterLink
-          class="px-3 py-1 rounded-full bg-foreground/10 text-sm ta cursor-pointer"
+          class="cursor-pointer rounded-full bg-foreground/10 px-3 py-1 text-sm ta-200 hover:bg-foreground/15"
           v-for="tag in item?.tags"
           :key="tag"
           :to="`/search?q=${tag}`"
@@ -38,7 +36,7 @@ const openUrl = (url: string | undefined) => {
           {{ tag }}
         </RouterLink>
       </div>
-      <vBtn @click="openUrl(item?.url)"> Visit </vBtn>
+      <vBtn @click="openUrl(`${item?.link}?ref=Awesome347`)"> Visit </vBtn>
     </div>
   </div>
 </template>