+++ src/components/vFoot.vue
@@ -4,7 +4,7 @@
       preserveAspectRatio="none"
       viewBox="0 0 1200 120"
       xmlns="http://www.w3.org/2000/svg"
-      class="fill-primary-600 w-[125%] h-[3.25rem] md:h-[4.6875rem] rotate-180"
+      class="h-[3.25rem] w-[125%] rotate-180 fill-primary-600 md:h-[4.6875rem]"
     >
       <path
         d="M0 0v46.29c47.79 22.2 103.59 32.17 158 28 70.36-5.37 136.33-33.31 206.8-37.5 73.84-4.36 147.54 16.88 218.2 35.26 69.27 18 138.3 24.88 209.4 13.08 36.15-6 69.85-17.84 104.45-29.34C989.49 25 1113-14.29 1200 52.47V0z"
@@ -19,12 +19,10 @@
       />
     </svg>
   </div>
-  <footer
-    class="bg-primary-600 text-primary-50 px-24 md:px-8 py-8 text-base md:text-lg grid place-items-center"
-  >
+  <footer class="grid place-items-center bg-primary-600 px-24 py-8 text-base text-primary-50 md:px-8 md:text-lg">
     <div>
       <p>Developed with ❤️ by <span class="font-bold">MFM-347</span></p>
-      <div class="font-semibold text-center">
+      <div class="text-center font-semibold">
         <a class="hover:underline" href="https://github.com/MFM-347/">Github</a>
         <span class="mx-4">/</span>
         <a class="hover:underline" href="mailto:madnifm347@outlook.com">Mail</a>
@@ -39,7 +37,7 @@
       </nav>
     </div>
     <div class="mt-8">
-      <p class="text-sm md:text-base font-medium">© 2025 MFM-347. All rights reserved.</p>
+      <p class="text-sm font-medium md:text-base">© 2025 MFM-347. All rights reserved.</p>
     </div>
   </footer>
 </template>
@@ -47,6 +45,6 @@
 @reference "../style.css";
 
 .lnk {
-  @apply border-x border-foreground px-4 py-2 hover:underline hover:font-semibold;
+  @apply border-x border-foreground px-4 py-2 ta-100 hover:font-semibold hover:underline;
 }
 </style>