+++ src/views/vItem.vue
@@ -3,7 +3,7 @@ import { storeToRefs } from 'pinia'
 import { useAwesomeStore } from '../stores'
 import { computed, onMounted } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
-import { GlobeAltIcon, CodeBracketIcon } from '@heroicons/vue/24/solid'
+import { GlobeAltIcon, CodeBracketIcon } from '@heroicons/vue/24/outline'
 
 const store = useAwesomeStore()
 const { items } = storeToRefs(store)
@@ -30,31 +30,29 @@ const openUrl = (url: string | undefined) => {
 
 <template>
   <div v-if="item" class="flex flex-col">
-    <div
-      v-if="item.icon"
-      class="relative top-2 mx-auto bg-yellow-400/40 font-semibold p-4 rounded-lg shadow-lg"
-    >
-      <b>Note:</b> Icons here are either favicons or from icon libraries. Ensure proper permissions
-      and licensing before use.
+    <div class="relative top-2 mx-auto rounded-lg bg-yellow-400/40 p-4 font-semibold shadow-lg">
+      <b>Note:</b> Icons here are either favicons or from icon libraries. Ensure proper permissions and licensing before
+      use.
     </div>
     <div class="flex justify-center p-6">
       <div
-        class="max-w-md w-full bg-card border rounded-xl shadow-lg text-card-foreground overflow-hidden"
+        class="bg-card text-card-foreground w-full max-w-md overflow-hidden rounded-xl border shadow-lg max-sm:w-[100%]"
       >
-        <div class="p-6 space-y-4">
+        <div class="space-y-4 p-6">
           <div class="flex items-center space-x-4">
-            <div v-if="item.icon" class="w-12 h-12 p-1 rounded-md bg-zinc-200">
-              <img :src="item.icon" :alt="`${item.name} Icon`" class="h-full w-full rounded-md" />
+            <div class="h-12 w-12 rounded-md bg-zinc-200 p-1">
+              <img v-if="item.icon" :src="item.icon" :alt="`${item.name} Icon`" class="h-full w-full rounded-md" />
+              <img v-else src="/ph.svg" alt="Placeholder Icon" class="h-full w-full rounded-md" />
             </div>
             <h3 class="text-4xl font-bold">{{ item.name }}</h3>
           </div>
-          <p class="text-xl mt-2"><b>Description:</b> {{ item.description }}</p>
+          <p class="mt-2 text-xl"><b>Description:</b> {{ item.description }}</p>
           <p class="text-lg"><b>Category:</b> {{ item.category }}</p>
           <p class="text-lg"><b>Open Source:</b> {{ item.foss }}</p>
-          <div class="flex flex-wrap gap-2 mt-2">
+          <div class="mt-2 flex flex-wrap gap-2">
             <span class="text-base"><b>Tags:</b>&nbsp;</span>
             <RouterLink
-              class="px-3 py-1 rounded-full bg-foreground/10 text-sm ta cursor-pointer"
+              class="cursor-pointer rounded-full bg-foreground/10 px-3 py-1 text-sm ta-100 hover:scale-105 hover:bg-foreground/20"
               v-for="tag in item?.tags"
               :key="tag"
               :to="`/search?q=${tag}`"
@@ -64,14 +62,14 @@ const openUrl = (url: string | undefined) => {
           </div>
           <h4 class="text-2xl font-semibold">Links:</h4>
           <div class="flex flex-wrap gap-2">
-            <button @click="openUrl(item?.url)">
+            <button @click="openUrl(`${item?.link}?ref=Awesome347`)">
               <GlobeAltIcon
-                class="size-8 fill-foreground hover:fill-primary-500 dark:hover:fill-primary-400 ta"
+                class="size-8 stroke-foreground ta-150 hover:stroke-primary-500 dark:hover:stroke-primary-400"
               /><span class="sr-only"> Check it out </span>
             </button>
-            <button v-if="item?.foss" @click="openUrl(item?.fossLnk)">
+            <button v-if="item?.foss" @click="openUrl(`${item?.link}?ref=Awesome347`)">
               <CodeBracketIcon
-                class="size-8 fill-foreground hover:fill-primary-500 dark:hover:fill-primary-400 ta"
+                class="size-8 stroke-foreground ta-150 hover:stroke-primary-500 dark:hover:stroke-primary-400"
               /><span class="sr-only"> Source Code </span>
             </button>
           </div>
@@ -80,11 +78,3 @@ const openUrl = (url: string | undefined) => {
     </div>
   </div>
 </template>
-
-<style scoped>
-@media (max-width: 640px) {
-  .max-w-md {
-    max-width: 100%;
-  }
-}
-</style>