+++ src/views/vNF.vue
@@ -4,9 +4,9 @@
     <p class="mb-10 text-xl sm:text-3xl">Sorry, we couldn't find this page.</p>
     <RouterLink
       to="/"
-      class="px-4 py-2 text-foreground bg-primary-500/70 hover:bg-primary-400/80 ta rounded-xl text-lg cursor-pointer"
+      class="cursor-pointer rounded-xl bg-primary-500/70 px-4 py-2 text-lg text-foreground ta-200 hover:bg-primary-400/80"
     >
-      Go home
+      Go Home
     </RouterLink>
   </section>
 </template>