+++ changelog.md
@@ -1,4 +1,4 @@
-# 1. Update 18-1-25
+# 1. Update 18-01-25
 
 **Commit Name:** `Upgrade list, UI and seo and add dynamic routes`
 
@@ -10,7 +10,7 @@
 4. Update list
 5. Add dynamic routes for each item in list
 
-# 2. Update 29-1-25
+# 2. Update 29-01-25
 
 **Commit Name:** `Upgrade tailwindcss and recreate UI`
 
@@ -18,9 +18,8 @@
 
 1. Recreated UI
 2. Upgraded to tailwindcss v4
-3. Updated to latest create-vue-template
 
-# 2. Update 30-1-25
+# 3. Update 30-01-25
 
 **Commit Name:** `Improve performance, type and clean-up SEO`
 
@@ -31,3 +30,15 @@
 3. Change typographic scheme from `roboto_condensed+source_sans_3` to `inter`
 4. Added `id` to data for consistent routing
 5. Added auto sitemap and robotsTxt generation
+
+# 4. Update 09-02-25
+
+**Commit Name:** `Upgraded features, UI, routing and list`
+
+**Changes:**
+
+1. Updated Layout
+2. Changed icons from `fill` to `outline`
+3. Upgraded search features, i.e, use x:Q where x =? ['name', 'category', 'tags', 'url', 'link', 'description']
+4. Updated list
+5. Changed dynamic routing from `/item/` to `/i/` for minimal URLs